Move location and package of model definitions
